[refactor] Moving twitteroauth lib to plugin Moving the library to the only extension that uses it. This should be replaced with the composer package but it has a php 7.2+ requirement.
Refs: https://packagist.org/packages/abraham/twitteroauth
