+++ core/plugins/authentication/twitter/twitteroauth/OAuth.php
@@ -1,0 +1,12 @@
+<?php
+include_once __DIR__ . '/TwitterOAuthException.php';
+include_once __DIR__ . '/TwitterOAuthConsumer.php';
+include_once __DIR__ . '/TwitterOAuthToken.php';
+include_once __DIR__ . '/TwitterOAuthSignatureMethod.php';
+include_once __DIR__ . '/TwitterOAuthSignatureMethod_HMAC_SHA1.php';
+include_once __DIR__ . '/TwitterOAuthSignatureMethod_PLAINTEXT.php';
+include_once __DIR__ . '/TwitterOAuthSignatureMethod_RSA_SHA1.php';
+include_once __DIR__ . '/TwitterOAuthRequest.php';
+include_once __DIR__ . '/TwitterOAuthServer.php';
+include_once __DIR__ . '/TwitterOAuthDataStore.php';
+include_once __DIR__ . '/TwitterOAuthUtil.php';
