+++ core/plugins/authentication/twitter/twitteroauth/TwitterOAuthException.php
@@ -1,0 +1,7 @@
+<?php
+/* Generic exception class
+ */
+class TwitterOAuthException extends Exception
+{
+	// pass
+}
